backend/requirements.txt
Django==1.10.6
gunicorn==19.7.1
djangorestframework==3.6.2
django-celery==3.2.1
psycopg2==2.7.1
redis==2.10.5
django-channels==0.7.0